Initial commit shows the problem
This commit is the initial commit introducing the issue with RedisTemplate.
In this commit, we added a test that preloads a key and value into redis. One test tries to fetch the key directly from redis and fails to find it, resulting in a `null` response. The other test, adding the key with a new value using RestTemplate, and loading it right after, this test is successful.

In the next commit, we could see that fixing the template will allow us to fetch the key also in the first test